Create macro for defining registers
Add registers-macro crate which defines a proc macro for generating
register definitions. This generates a RegisterId enumerator with
a constructor for each named register. It also generates a static
array of RegisterInfo instances which describe each register in
the system.

The macro is used within the librsdb::register module where the
definitions will be generated.