Add function to write memory at the given address
Add Process::write_memory which writes data from the given slice
a word at a time.

Add TryFromBytes trait which tries to read a type from a byte
slice. This is a safe counterpart to the FromBytesRaw trait.
Implement this in the existing macro and rename it to
derive_from_bytes.