Create syscalls! macro to generate SyscallType enum
Create a syscalls! macro in the existing registers-macro crate
which takes a sequence of syscall (name, id) pairs and uses them
to generate a SyscallType enum.

Add gen_syscalls.awk script to generate the required call to this
macro from the unistd.h header file in the linux includes. Include
this file within the new syscalls module to generate the SyscallType
enum.