Add catchpoint command
Add catchpoint command which allows syscalls to be trapped on entry
and exit.

Extend the syscalls! macro to generate implementations of FromStr
and a from_id method.

Change type of tracked syscalls to SyscallType instead of their ids.
Update the SyscallInfo type to use this instead of the raw id.