Resolve function name when stopped
Try to resolve the name of the function when displaying the stop
reason.

Remove Display instance from StopReason and create a new
TargetStopReason enum which can be displayed instead. This contains
the name of the function when stopped if it can be resolved. This
information is only available from the debugger and not the raw
process.