Make DIE and attribute offsets relative to start of .debug_info
The current offsets calculated for DIEs and their attributes
are relative to the start of the data segment of their parent
compilation unit (i.e. directly after the compile unit header).
Change these to be offsets from the start of the .debug_info
section. This means attributes values can be resolved directly
without reference to the parent compile unit. One exception is
reference attributes which may point to an offset within the
parent compile unit.