Add id type for compile unit and add to DIE
Add CompileUnitId type which identifies a compile unit within
the DWARF info of an ELF file.

Store compile units in a map keyed by id for faster lookup.

Add id for the parent compile unit to DIE. This allows the Abbrev
entry to be resolved without requiring the parent compile unit to
be passed in as a parameter.